style : 회원가입 페이지 수정 ref : #17
